08.05.2023 - тесты для приватного и админа готовы

+++ ewm-main-service/src/main/java/ru/practicum/events/comments/service/impl/CommentsServicePrivateImpl.java
@@ -86,8 +86,9 @@ public class CommentsServicePrivateImpl implements CommentsServicePrivate {
 
     @Override
     public void deleteComment(Long commentId, Long userId) {
-        findObjectInRepository.getUserById(userId);
+        User user = findObjectInRepository.getUserById(userId);
         Comment comment = findObjectInRepository.getCommentById(commentId);
+        processingComment.checkCommentOnOwner(comment, user);
         commentsRepository.delete(comment);
     }
